refactor(plugin): move tool implementation to plugin
