+++ packages/plugin/tsconfig.json
@@ -5,7 +5,8 @@
     "rootDir": "src",
     "outDir": "dist",
     "declaration": true,
-    "lib": ["es2022", "dom", "dom.iterable"]
+    "lib": ["ESNext", "DOM", "DOM.Iterable"],
+    "noUncheckedIndexedAccess": false
   },
   "include": ["src"]
 }
