+++ packages/sdk-next/src/tool.ts
@@ -1,2 +1,2 @@
-export { Failure, RegistrationError, make } from "@opencode-ai/core/tool/tool"
-export type { AnyTool, Content, Context, Definition } from "@opencode-ai/core/tool/tool"
+export { Failure, RegistrationError, make } from "@opencode-ai/plugin/v2/effect/tool"
+export type { AnyTool, Content, Context, Definition } from "@opencode-ai/plugin/v2/effect/tool"
